Fix float64/float32 dtype mismatch after scipy sosfilt
scipy.signal.sosfilt promotes output to float64 regardless of input
dtype. torch.from_numpy then creates a float64 tensor which clashes
with the float32 mel_fb buffer in forward_frame.

Fix: cast back to float32 after sosfilt in run_inference.py (offline
eval) and live.py (hardware inference).

+++ eval/run_inference.py
@@ -100,8 +100,8 @@ def run_batch(val_dir=None, out_dir=None, checkpoint=None):
         # Console HPF — apply to the full file before the frame loop (stateful, causal).
         # mic: matches training (HPF applied to mic in recursive_train.py).
         # ref: HPF'd here because training teacher-forcing ref is HPF'd reverberant vocal.
-        mic_np = sosfilt(_console_hpf, mic_np)
-        ref_np = sosfilt(_console_hpf, ref_np)
+        mic_np = sosfilt(_console_hpf, mic_np).astype(np.float32)
+        ref_np = sosfilt(_console_hpf, ref_np).astype(np.float32)
 
         # Pad to a whole number of HOP blocks so the last partial frame is not silently dropped.
         pad      = (-min_len) % HOP
